+++ tests/corpus/raw-baseline/arm64_O1/pearson.c
@@ -14,7 +14,7 @@ uint64_t sym__pearson(uint64_t X0_0, uint64_t X1_0)
                 X0_0++;
                 uint8_t tmp_25400_2 = *tmp_7400_2;
                 uint8_t tmp_25600_2 = *(uint8_t*)(((uint64_t)((uint32_t)tmp_25400_2 ^ (uint32_t)X8_1) & 255) + X9_2);
-                X8_1 = (uint64_t)(uint8_t)tmp_25600_2;
+                X8_1 = (uint64_t)tmp_25600_2;
                 uint8_t TMPZR_3 = X1_0 == 1;
                 X1_0--;
                 if (TMPZR_3) {
