+++ tests/corpus/raw-baseline/arm64_O1/unaligned_words.c
@@ -13,7 +13,7 @@ uint64_t sym__unaligned_words(uint64_t X0_0, uint64_t X1_0)
             uint64_t X10_1 = (uint64_t)X8_1 + 1;
             for (; ; ) {
                 uint32_t tmp_25180_2 = *(uint32_t*)(X12_1 + X10_1);
-                X0_2 = (uint64_t)(uint32_t)(tmp_2a000_4 * ((uint32_t)X0_2 ^ tmp_25180_2));
+                X0_2 = (uint64_t)(tmp_2a000_4 * ((uint32_t)X0_2 ^ tmp_25180_2));
                 X11_1 = X12_1 + 7;
                 uint64_t X13_2 = X12_1 + 15;
                 X12_1 = X11_1;
@@ -34,7 +34,7 @@ uint64_t sym__unaligned_words(uint64_t X0_0, uint64_t X1_0)
                     uint8_t* tmp_7400_2 = X8_2;
                     X8_2 = (uint8_t*)((uint64_t)X8_2 + 1);
                     uint8_t tmp_25400_2 = *tmp_7400_2;
-                    X0_2 = (uint64_t)(uint32_t)(((uint32_t)X0_2 ^ (uint32_t)tmp_25400_2) * tmp_2a000_4);
+                    X0_2 = (uint64_t)(((uint32_t)X0_2 ^ (uint32_t)tmp_25400_2) * tmp_2a000_4);
                     uint8_t TMPZR_11 = X10_3 == 1;
                     X10_3--;
                     if (TMPZR_11) {
