+++ tests/corpus/raw-baseline/arm64_O2/sdbm.c
@@ -13,7 +13,7 @@ uint64_t sym__sdbm(uint64_t X0_0, uint64_t X1_0)
             uint8_t* tmp_7400_2 = X8_1;
             X8_1 = (uint8_t*)((uint64_t)X8_1 + 1);
             uint8_t tmp_25400_2 = *tmp_7400_2;
-            X0_1 = (uint64_t)(uint32_t)((uint32_t)X0_1 * tmp_2a000_2 + (uint32_t)tmp_25400_2);
+            X0_1 = (uint64_t)((uint32_t)X0_1 * tmp_2a000_2 + (uint32_t)tmp_25400_2);
             uint8_t TMPZR_2 = X1_0 == 1;
             X1_0--;
             if (TMPZR_2) {
