+++ tests/corpus/raw-baseline/x64_O1/xxhash32.c
@@ -33,7 +33,7 @@ uint32_t sym__xxhash32(uint64_t RDI_0, uint64_t RSI_0, uint64_t RDX_0)
                 R9_1 = (uint32_t)tmp_lane_100000910_91_1c_1;
                 uint32_t tmp_lane_100000910_99_1f_1 = (uint32_t)((uint64_t)tmp_11f00_5 * 0xffffffff85ebca77U) + (uint32_t)RCX_1;
                 tmp_lane_100000910_b8_24_1 = (uint32_t)((uint64_t)(int32_t)(tmp_lane_100000910_67_16_1 << 13 | tmp_lane_100000910_67_16_1 >> 19) * 0xffffffff9e3779b1U);
-                RDX_0 = (uint64_t)(uint32_t)tmp_lane_100000910_b8_24_1;
+                RDX_0 = (uint64_t)tmp_lane_100000910_b8_24_1;
                 tmp_lane_100000910_c1_26_1 = (uint32_t)((uint64_t)(int32_t)(tmp_lane_100000910_99_1f_1 << 13 | tmp_lane_100000910_99_1f_1 >> 19) * 0xffffffff9e3779b1U);
                 RCX_1 = (uint32_t)tmp_lane_100000910_c1_26_1;
                 RDI_0 += 16;
